refactor version check
